refactor: refactored firebase config file and location
