+++ tsconfig.json
@@ -21,6 +21,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "cypress"],
+  "include": ["src", "cypress", "firebase.config.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
